Refactor containers
